+++ Editor/Discover/DiscoverWindow.cs
@@ -85,7 +85,7 @@ namespace GameplayIngredients.Editor
         {
             if (discoverAsset != null)
             {
-                var window = GetWindow<DiscoverWindow>(true);
+                var window = GetWindow<DiscoverWindow>(!discoverAsset.dockable);
                 window.SetDiscoverAsset(discoverAsset);
             }
             else
@@ -150,6 +150,34 @@ namespace GameplayIngredients.Editor
             UpdateDiscoverObjects();
         }
 
+
+        /// <summary>
+        /// Filters object based on asset tags, object tags and FilterMode
+        /// </summary>
+        /// <param name="asset"></param>
+        /// <param name="discoverObject"></param>
+        /// <returns>true to discard, false to keep</returns>
+        static bool FilterDiscoverObject(DiscoverAsset asset, Discover discoverObject)
+        {
+            if (string.IsNullOrEmpty(asset.Tags))
+                asset.Tags = string.Empty;
+            if (string.IsNullOrEmpty(discoverObject.Tags))
+                discoverObject.Tags = string.Empty;
+
+            string[] assetTags = asset.Tags.Split(' ');
+            string[] objectTags = discoverObject.Tags.Split(' ');
+
+            switch(asset.filterMode)
+            {
+                case DiscoverAsset.FilterMode.ShowAll: return false;
+                case DiscoverAsset.FilterMode.IncludeTags:
+                    return !assetTags.Intersect(objectTags).Any();
+                case DiscoverAsset.FilterMode.ExcludeTags:
+                    return assetTags.Intersect(objectTags).Any();
+            }
+            return false;
+        }
+
         void UpdateDiscoverObjects(bool clear = false)
         {
             if (discoverObjects == null)
@@ -163,6 +191,10 @@ namespace GameplayIngredients.Editor
             // Add new ones
             foreach (var item in newOnes)
             {
+                // Apply Filter
+                if (FilterDiscoverObject(discoverAsset, item))
+                    continue;
+
                 if (!discoverObjects.ContainsKey(item.Category))
                 {
                     discoverObjects.Add(item.Category, new List<Discover>());
@@ -318,6 +350,11 @@ namespace GameplayIngredients.Editor
                 GUILayout.Label(discoverAsset.Title, Styles.header);
                 using (new GUILayout.VerticalScope(Styles.indent))
                 {
+                    if(discoverAsset.Image != null)
+                    {
+                        DiscoverEditor.DrawImage(discoverAsset.Image);
+                    }
+
                     GUILayout.Label(discoverAsset.Description, Styles.body);
 
                     if(discoverAsset.Scenes != null)
@@ -326,6 +363,11 @@ namespace GameplayIngredients.Editor
                         {
                             using (new GroupLabelScope(map.Title))
                             {
+                                if(map.Image != null)
+                                {
+                                    DiscoverEditor.DrawImage(map.Image);
+                                }
+
                                 GUILayout.Label(map.Description, Styles.body);
 
                                 using (new GUILayout.HorizontalScope())
@@ -553,6 +595,8 @@ namespace GameplayIngredients.Editor
 
             public static GUIStyle tabContainer;
 
+            public static GUIStyle image;
+
             static Styles()
             {
                 header = new GUIStyle(EditorStyles.wordWrappedLabel);
@@ -578,8 +622,8 @@ namespace GameplayIngredients.Editor
 
                 boxHeader = new GUIStyle(GUI.skin.box);
                 boxHeader.normal.textColor = GUI.skin.label.normal.textColor;
-                boxHeader.fixedHeight = 20;
-                boxHeader.fontSize = 11;
+                boxHeader.fixedHeight = 24;
+                boxHeader.fontSize = 16;
                 boxHeader.fontStyle = FontStyle.Bold;
                 boxHeader.alignment = TextAnchor.UpperLeft;
                 boxHeader.margin = new RectOffset(0, 0, 0, 6);
@@ -596,6 +640,9 @@ namespace GameplayIngredients.Editor
 
                 tabContainer = new GUIStyle(EditorStyles.miniButton);
                 tabContainer.padding = new RectOffset(4, 4, 0, 0);
+
+                image = new GUIStyle(GUIStyle.none);
+                image.stretchWidth = true ;
 
             }
         }