+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2019.3.5
+
+#### Added
+
+- Discover : Tag Filtering System
+- Discover : Images in Discover Window
+
 ## 2019.3.4
 
 This version added the following package dependencies: